Ref: Covid Requests To DRY && Clean Daily Checks
I read your article and was inspired to do the same for my wife.
I was also curious to your approach. So, I hope you don't mind me taking the liberty to assist. 

There isn't much here, mainly: 
1. a refactor of the outbound network request
2. a spread of the Points object into our Daily results container.
3. a simple dailyCase check.

I left comments in the code for you. Feel free to reject but, thanks for considering.
I admit though, I don't recall the response of Cheerio if there is an internal error with load.
You may want to guard against that. But, if load throws, you'll be ok since it'll be caught and null returned from finally.